refactor users fetch
