+++ src/services/usersService.js
@@ -1,42 +1,33 @@
-import { MAIN_URL } from '../../constants';
+import { MAIN_URL } from '../constants';
+import { getHeaders } from './utils';
 
 export const getUserInvitationRequestData = ({ token, body }) =>
   new Request(`${MAIN_URL}auth/invite`, {
     method: 'POST',
-    headers: {
-      'Content-Type': 'application/json;charset=utf-8',
-      Authorization: `Bearer ${token}`,
-    },
+    headers: getHeaders(token),
     body: JSON.stringify(body),
   });
 
 export const getUserUpdateRequestData = ({ token, id, body }) =>
   new Request(`${MAIN_URL}users/${id}`, {
     method: 'PUT',
-    headers: {
-      'Content-Type': 'application/json;charset=utf-8',
-      Authorization: `Bearer ${token}`,
-    },
+    headers: getHeaders(token),
     body: JSON.stringify(body),
+  });
+
+export const getUserDeleteRequestData = ({ token, id }) =>
+  new Request(`${MAIN_URL}users/${id}`, {
+    method: 'DELETE',
+    headers: getHeaders(token),
   });
 
 export const getLimitedUsersRequestData = (token, limit, page, isAdmin) =>
   new Request(
     `${MAIN_URL}users${isAdmin ? `/all?limit=${limit}&page=${page}` : ''}`,
     {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
+      headers: getHeaders(token),
     }
   );
-
-export const getUserDeleteRequestData = ({ token, id }) =>
-  new Request(`${MAIN_URL}users/${id}`, {
-    method: 'DELETE',
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
 
 export const getFilteredUsersRequestData = (
   token,
@@ -50,8 +41,6 @@ export const getFilteredUsersRequestData = (
       value ? `&first_name=${value}` : ''
     }&limit=${limit}&page=${page}`,
     {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
+      headers: getHeaders(token),
     }
   );
